adding tests module
